Move persistence helper functions into class Persistence
